+++ src/libs/server/notion.ts
@@ -270,6 +270,7 @@ export const getNotionPreviewImages = async recordMap => {
     },
     {
       concurrency: notionConfig.previeImages.concurrency,
+      stopOnError: false,
     }
   )