+++ src/libs/server/page.ts
@@ -45,6 +45,8 @@ export const showCommonPage = (
         hasError: true,
       },
     }
+    // trigger rewrite to failsafe page
+    res.setHeader('x-dazedbear-failsafe', 1)
   }
 
   const options: logOption = {
@@ -93,7 +95,7 @@ export const fetchArticleStream = async ({
   pageId,
   category,
 }: {
-  req: GetServerSidePropsRequest
+  req?: GetServerSidePropsRequest
   pageName?: NotionPageName
   pageId?: string
   category?: string