+++ src/libs/server/transformer.ts
@@ -61,19 +61,6 @@ export const transformArticleStream = async (
     ? articleStream.ids.length - 1
     : 0
 
-  const isPreviewImageGenerationEnabled: boolean = get(notion, [
-    'previeImages',
-    'enable',
-  ])
-
-  if (!isEmpty(articleStream.content) && isPreviewImageGenerationEnabled) {
-    // we only parse the cover preview images from the recordMap of collection data to prevent duplicated parsing from blog/[slug]
-    const previewImageMap: PreviewImagesMap = await getNotionPreviewImages(
-      articleStream.content
-    )
-    set(articleStream, ['content', 'preview_images'], previewImageMap)
-  }
-
   const schema = {
     type: 'object',
     additionalProperties: false,
@@ -108,6 +95,51 @@ export const transformArticleStream = async (
     }
     log(options)
     throw 'articleStream is invalid'
+  }
+
+  return articleStream
+}
+
+export const transformArticleStreamPreviewImages = async (
+  articleStream: ArticleStream
+): Promise<ArticleStream> => {
+  const isPreviewImageGenerationEnabled: boolean = get(notion, [
+    'previeImages',
+    'enable',
+  ])
+
+  if (!isEmpty(articleStream.content) && isPreviewImageGenerationEnabled) {
+    // we only parse the cover preview images from the recordMap of collection data to prevent duplicated parsing from blog/[slug]
+    const previewImageMap: PreviewImagesMap = await getNotionPreviewImages(
+      articleStream.content
+    )
+    set(articleStream, ['content', 'preview_images'], previewImageMap)
+
+    const schema = {
+      type: 'object',
+      additionalProperties: true,
+      properties: {
+        content: {
+          type: 'object',
+          properties: {
+            preview_images: {
+              type: 'object',
+            },
+          },
+          required: ['preview_images'],
+        },
+      },
+      required: ['content'],
+    }
+    const validate = ajv.compile(schema)
+    if (!validate(articleStream)) {
+      const options: logOption = {
+        category: 'transformArticleStreamPreviewImages',
+        message: validate.errors,
+        level: 'warn',
+      }
+      log(options)
+    }
   }
 
   return articleStream
@@ -226,7 +258,7 @@ export const transformPageUrls = (
   articleStream: ArticleStream
 ): string[] => {
   const recordMap = articleStream.content
-  const pageUrls: string[] = articleStream.ids.map(pageId => {
+  let pageUrls: string[] = articleStream.ids.map(pageId => {
     const pagePath = getSinglePagePath({
       pageName,
       pageId,
@@ -234,6 +266,8 @@ export const transformPageUrls = (
     })
     return `/${pageName}/${pagePath}`
   })
+
+  pageUrls.push(`/${pageName}`)
 
   const schema = {
     type: 'array',