+++ src/pages/api/sitemap.ts
@@ -12,18 +12,24 @@ import {
   transformArticleStream,
   transformPageUrls,
 } from '../../libs/server/transformer'
-import { navigation, notion, cache as cacheConfig } from '../../../site.config'
+import {
+  pages,
+  notion,
+  cache as cacheConfig,
+  website,
+} from '../../../site.config'
 import cacheClient from '../../libs/server/cache'
 
 const route = '/sitemap'
 const methods = ['GET']
+const currentEnv = process.env.NEXT_PUBLIC_APP_ENV || 'production'
 
 dayjs.extend(utc)
 const category = getCategory(route)
 
 const generateSiteMapXml = async req => {
   // get all enabled static page paths
-  const pageUrls = navigation
+  const pageUrls = Object.values(pages)
     .map(item => item.enabled && item.page)
     .filter(path => path)
 
@@ -50,6 +56,7 @@ const generateSiteMapXml = async req => {
     },
     {
       concurrency: 10,
+      stopOnError: false,
     }
   )
   let notionUrls: string[] = currentNotionListUrls.reduce(
@@ -58,10 +65,15 @@ const generateSiteMapXml = async req => {
   )
 
   // all collected urls
-  const urls = ['/'].concat(pageUrls, notionUrls).map(url => ({ url }))
+  const urls = [].concat(pageUrls, notionUrls).map(url => ({ url }))
 
   // generate sitemap xml
-  const stream = new SitemapStream({ hostname: `https://${req.headers.host}` })
+  const stream = new SitemapStream({
+    hostname: `${get(website, [currentEnv, 'protocol'])}://${get(website, [
+      currentEnv,
+      'host',
+    ])}`,
+  })
   const sitemapXml = await streamToPromise(
     Readable.from(urls).pipe(stream)
   ).then(data => data.toString())